Base classes moved into their child folders
