+++ apis/core-api/src/services/BaseService.ts
